pyproject.toml
[tool.poetry]
name = "vadis_summarization_api"
version = "0.1.0"
description = ""
authors = ["sobamchan <oh.sore.sore.soutarou@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
python-dotenv = "^0.20.0"
torch = "^1.11.0"
transformers = "^4.17.0"
sienna = "^0.1.5"
langdetect = "^1.0.9"
sacremoses = "^0.0.53"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
neovim = "^0.3.1"
ipython = "^8.2.0"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
